add initial sf-level terra run integration tests
