+++ tests/integration/step-function/mock_sf_cfg.json
@@ -58,6 +58,17 @@
         "RequestApprovalFails": {
           "Plan": "MockedEcsSuccess",
           "Request Approval": "MockedApprovalRequestFails"
+        },
+        "TestApply": {
+          "Plan": "MockedEcsSuccess",
+          "Request Approval": "MockedApprove"
+        },
+        "TestApprovalRequest": {
+          "Plan": "MockedEcsSuccess"
+        },
+        "TestPlan": {
+          "Apply": "MockedEcsSuccess",
+          "Request Approval": "MockedApprove"
         }
       }
     }
